backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
